perf: use other algo replace crypto uuid
