+++ src/utils/time.js
@@ -114,3 +114,20 @@ export function getDayFuture(days, now) {
 export function sleep(time) {
   return new Promise((resolve) => setTimeout(resolve, time))
 }
+
+export function formatDate(inputTime) {
+  const date = new Date(inputTime)
+  const y = date.getFullYear()
+  let m = date.getMonth() + 1
+  m = m < 10 ? ('0' + m) : m
+  let d = date.getDate()
+  d = d < 10 ? ('0' + d) : d
+  let h = date.getHours()
+  h = h < 10 ? ('0' + h) : h
+  let minute = date.getMinutes()
+  let second = date.getSeconds()
+  minute = minute < 10 ? ('0' + minute) : minute
+  second = second < 10 ? ('0' + second) : second
+  // return y + '-' + m + '-' + d + ' ' + h + ':' + minute + ':' + second
+  return y + '-' + m + '-' + d + 'T' + h + ':' + minute + ':' + second
+}
